Session tests now check expiration dates with second precision.
When comparing expiration dates in the various test cases, the dates
are checked against the full string representation. This represents
an issue, due to the nature of the tests (specially when a new request
needs to be performed, computation might take more than 1 second, depending
on the current system load).

This commit makes sure that the dates are checked with second precision, to
account for any delay running the tests because of system load while running
the tests.